+++ src/extension.ts
@@ -605,8 +605,8 @@ function createNewEditorInstance(context: vscode.ExtensionContext, activeTextEdi
 
 	panel.webview.html = createEditorHtml(panel.webview, context, config, {
 		isWatchingSourceFile: instance.supportsAutoReload,
-		sourceFileCursorLineIndex: config.openTableAtCursorPos === 'onlyInitially' ? activeTextEditor.selection.active.line : null,
-		sourceFileCursorColumnIndex: config.openTableAtCursorPos === 'onlyInitially' ? activeCol : null,
+		sourceFileCursorLineIndex: config.openTableAndSelectCellAtCursorPos === 'initialOnly' ? activeTextEditor.selection.active.line : null,
+		sourceFileCursorColumnIndex: config.openTableAndSelectCellAtCursorPos === 'initialOnly' ? activeCol : null,
 		isCursorPosAfterLastColumn: activeTextEditor.document.lineAt(activeTextEditor.selection.active.line).text.length === activeTextEditor.selection.active.character
 	})